refactor(suite-native): extract wallet label package
