+++ suite-native/address/README.md
@@ -1,0 +1,8 @@
+### `@suite-native/address`
+
+This package is supposed to only contain address related code.
+address Labeling, components to display the address itself,
+selectors for the address, ... are all ok.
+
+But there shall be no trading, account, etc... On the contrary,
+this package should be used in such packages.
