+++ suite-native/device-manager/src/components/WalletItemBase.tsx
@@ -13,7 +13,8 @@ import { HStack, Radio, Text } from '@suite-native/atoms';
 import { BaseCurrencyAmountFormatter } from '@suite-native/formatters';
 import { Icon } from '@suite-native/icons';
 import { Translation } from '@suite-native/intl';
-import { WalletLabel, selectIsLabellingAllowed } from '@suite-native/labeling';
+import { selectIsLabellingAllowed } from '@suite-native/labeling';
+import { WalletLabel } from '@suite-native/wallet';
 import { prepareNativeStyle, useNativeStyles } from '@trezor/styles-native';
 
 type WalletItemBaseVariant = 'standard' | 'passphrase';
