+++ suite-native/labeling/README.md
@@ -1,0 +1,4 @@
+### `@suite-native/labeling`
+
+This package MUST contain only labeling code, agnostic to the what is labeled.
+There MUST not be any address, account, wallet, ...code.
