+++ suite-native/labeling/src/index.ts
@@ -1,5 +1,4 @@
-export { WalletLabel } from './components/WalletLabel';
-export { EditableLabelLayout } from './components/EditableLabelLayout';
-export { LabelEditForm } from './components/LabelEditForm';
+export { EditableLabelLayout } from './EditableLabelLayout';
+export { LabelEditForm } from './LabelEditForm';
 export type { CombinedLabelingState } from './selectors';
 export { selectIsLabellingAllowed } from './selectors';
