+++ suite-native/module-settings/package.json
@@ -43,7 +43,6 @@
         "@suite-native/helpers": "workspace:*",
         "@suite-native/icons": "workspace:*",
         "@suite-native/intl": "workspace:*",
-        "@suite-native/labeling": "workspace:*",
         "@suite-native/link": "workspace:*",
         "@suite-native/module-connect-popup": "workspace:*",
         "@suite-native/navigation": "workspace:*",
@@ -55,6 +54,7 @@
         "@suite-native/toasts": "workspace:*",
         "@suite-native/trading-residence": "workspace:*",
         "@suite-native/trading-state": "workspace:*",
+        "@suite-native/wallet": "workspace:*",
         "@trezor/blockchain-link-types": "workspace:*",
         "@trezor/connect": "workspace:*",
         "@trezor/env-utils": "workspace:*",
