+++ suite-native/wallet/README.md
@@ -1,0 +1,8 @@
+### `@suite-native/wallet`
+
+This package is supposed to only contain wallet related code.
+Wallet Labeling, components to display the wallet itself,
+selectors for the wallet, ... are all ok.
+
+But there shall be no trading, account, etc... On the contrary,
+this package should be used in such packages.
