+++ suite-native/wallet/package.json
@@ -1,0 +1,22 @@
+{
+    "name": "@suite-native/wallet",
+    "version": "1.0.0",
+    "private": true,
+    "license": "See LICENSE.md in repo root",
+    "sideEffects": false,
+    "main": "src/index",
+    "scripts": {
+        "depcheck": "yarn g:depcheck",
+        "type-check": "yarn g:tsc --build",
+        "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'"
+    },
+    "dependencies": {
+        "@reduxjs/toolkit": "2.11.2",
+        "@suite-common/suite-sync": "workspace:*",
+        "@suite-common/wallet-utils": "workspace:*",
+        "@suite-native/labeling": "workspace:*",
+        "@trezor/connect": "workspace:*",
+        "react": "19.2.4",
+        "react-redux": "9.2.0"
+    }
+}
