+++ suite-native/wallet/src/index.ts
@@ -1,0 +1,1 @@
+export { WalletLabel } from './WalletLabel';
